Add setting to adjust HMC API request timeout
When hosts have lost their connection to an HMC, the Managed System(s)
APIs will take 2+ minutes to respond. The default response timeout for
the rest-client gem is 60 seconds. Adding a user configurable setting to
override the API request timeout allows inventory refresh to succeed,
even when the responses are slow due to disconnected systems (or
possible other reasons).